Slighly refactored architecture to help others learn the app faster and improve scaling later on
