+++ package.json
@@ -19,8 +19,8 @@
     "web-vitals": "^1.1.0"
   },
   "scripts": {
-    "build:tailwind": "tailwind build src/styles/tailwind.css -o src/styles/main.css",
-    "watch:tailwind": " chokidar 'src/**/*.css' 'src/**/*.scss' --ignore src/styles/main.css -c 'npm build:tailwind' ",
+    "build:tailwind": "tailwind build src/ui/styles/tailwind.css -o src/ui/styles/main.css",
+    "watch:tailwind": " chokidar 'src/**/*.css' 'src/**/*.scss' --ignore src/ui/styles/main.css -c 'npm build:tailwind' ",
     "start": "npm-run-all build:tailwind --parallel watch:tailwind start:react",
     "start:react": "react-scripts start",
     "prebuild": "npm run build:tailwind",
