+++ src/logic/helperFunctions.js
