+++ src/ui/assets/icons.js
