+++ src/ui/assets/photos/breakfastTablePhoto.png
