+++ src/ui/views/landingPage/components/EmailNotification.jsx
