+++ src/ui/views/landingPage/components/LocationAsset.jsx
