+++ src/ui/views/landingPage/components/SearchAnimationComponents/FauxSearchBar.jsx
